wip: separate folder traversal state
